mergedish z buffer

+++ src/window.rs
@@ -5,7 +5,7 @@ pub struct Window {
     pub width: usize,
     pub height: usize,
     pub buffer: Vec<u32>, // encoding: 0RGB
-    pub depth_buffer: Vec<usize>,
+    pub depth_buffer: Vec<f32>,
 }
 impl Window {
     pub fn new(scale: usize, width: usize, height: usize) -> Self {
@@ -32,7 +32,7 @@ impl Window {
             width,
             height,
             buffer: vec![0; width * height],
-            depth_buffer: vec![usize::MAX; width * height],
+            depth_buffer: vec![f32::MAX; width * height],
         }
     }
